src/requirements.txt
jupyter>=1.0.0, <2.0
pcloud
click
PyYAML